% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{build.topology.subset}
\alias{build.topology.subset}
\title{Compute subset relation as edge list}
\usage{
build.topology.subset(samples)
}
\arguments{
\item{samples}{input dataset (mutational matrix) as matrix}
}
\value{
the computed edge list
}
\description{
Create an edge list E representing the 'subset' relation for binary strings so that:
\deqn{ (A,B) in E <=> forall(i) : A[i] -> B[i] }
}
\examples{
# compact dataset
compactedDataset <- compact.dataset.easy(dataset)
# turn it to matricial form
samples <- as.matrix(compactedDataset \%>\% select(-freq))
# compute topology
build.topology.subset(samples)

}
