% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMICE.R
\docType{package}
\name{CIMICE}
\alias{CIMICE}
\title{CIMICE Package}
\description{
R implementation of the CIMICE tool.
CIMICE is a tool in the field of tumor phylogenetics and
its goal is to build a Markov Chain (called Cancer Progression Markov Chain, CPMC) in order to model tumor subtypes evolution.
The input of CIMICE is a Mutational Matrix, so a boolean matrix representing altered genes in a
collection of samples. These samples are assumed to be obtained with single-cell DNA analysis techniques and
the tool is specifically written to use the peculiarities of this data for the CMPC construction.
See `https://github.com/redsnic/tumorEvolutionWithMarkovChains/tree/master/GenotypeEvolutionPaths` for the
original Java version of this tool.
}
\details{
CIMICE-R: (Markov) Chain Method to Infer Cancer Evolution
}
\author{
Nicolò Rossi \email{olocin.issor@gmail.com}
}
