% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{computeUPW.aux}
\alias{computeUPW.aux}
\title{Up weights computation (aux)}
\usage{
computeUPW.aux(g, edge, freqs, no.of.children, A)
}
\arguments{
\item{g}{graph (a Directed Acyclic Graph)}

\item{edge}{the currently considered edge}

\item{freqs}{observed genotype frequencies}

\item{no.of.children}{number of children for each node}

\item{A}{adjacency matrix of G}
}
\value{
a vector containing the Up weights for each edge
}
\description{
Computes the up weights formula:
}
\details{
\deqn{ W_{up}(<a,b> in E) = (1/|L_a|)(P(a) + \Sigma{for: x in \Pi_a}{of: W_{up}(<x,a>))}}

using a Dinamic Programming approach (recursion), see vignettes for further explaination.
}
\examples{
computeUPW.aux(g, edge, freqs, no.of.children, A)

}
