% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{corrplot.from.df}
\alias{corrplot.from.df}
\title{Dataframe based correlation plot}
\usage{
corrplot.from.df(df)
}
\arguments{
\item{df}{input dataset}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a dataframe
}
\examples{
corrplot.from.df(df)

}
