% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{corrplot.samples}
\alias{corrplot.samples}
\title{Sample based correlation plot}
\usage{
corrplot.samples(df)
}
\arguments{
\item{df}{input dataset (mutational matrix)}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a dataframe computed from samples' perspective
}
\examples{
corrplot.samples(df)

}
