% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dataset.preprocessing.population}
\alias{dataset.preprocessing.population}
\title{Run CIMICE preprocessing for poulation format dataset}
\usage{
dataset.preprocessing.population(dataset)
}
\arguments{
\item{dataset}{a mutational matrix as a dataframe (with freq column)}
}
\value{
a list containing the mutational matrix ("samples"),
the mutational frequencies of the genotypes ("freqs"),
the node labels ("labels") and finally the gene names ("genes")
}
\description{
executes the preprocessing steps of CIMICE
}
\details{
Preprocessing steps:

1) genotype frequencies are computed

2) labels are prepared
}
\examples{
require(dplyr)
example.dataset() \%>\% dataset.preprocessing.population

}
