% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw.ggraph}
\alias{draw.ggraph}
\title{Ggplot graph output}
\usage{
draw.ggraph(g, W, labels, digits = 4)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}

\item{digits}{precision for edges' weights}
}
\value{
ggraph object representing g as described
}
\description{
Draws the output graph using ggplot
}
\examples{
draw.ggraph(g,W,labels,digit = 3)

}
