% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw.visNetwork}
\alias{draw.visNetwork}
\title{VisNetwork graph output (default)}
\usage{
draw.visNetwork(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
visNetwork object representing g as described
}
\description{
Draws the output graph using VisNetwork
}
\examples{
draw.visNetwork(g,W,labels)

}
