% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{example.dataset.withFreqs}
\alias{example.dataset.withFreqs}
\title{Creates a simple example dataset with frequency column}
\usage{
example.dataset.withFreqs()
}
\value{
a simple mutational matrix
}
\description{
Creates a simple example dataset with frequency column
}
\examples{
example.dataset.withFreqs()

}
