% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{get.no.of.children}
\alias{get.no.of.children}
\title{Get number of children}
\usage{
get.no.of.children(A, g)
}
\arguments{
\item{A}{Adjacency matrix of the graph g}

\item{g}{a graph}
}
\value{
a vector containing the number of children for each node in g
}
\description{
Compute number of children for each node given an adj matrix
}
\examples{
A <- as.matrix(as_adj(g))
get.no.of.children(A, g)

}
