% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetGenerator.R
\name{make_generator_stub}
\alias{make_generator_stub}
\title{Create a stub for a generator}
\usage{
make_generator_stub(dataset)
}
\arguments{
\item{dataset}{A compacted CIMICE dataset}
}
\value{
a generator, with weight = 0 for all the edges
}
\description{
Create a generator topology directly from a dataset.
The topology will follow the subset relation.
}
\examples{
make_generator_stub(example_dataset())

}
