% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{normalizeDWNW}
\alias{normalizeDWNW}
\title{Down weights normalization}
\usage{
normalizeDWNW(g, freqs, no.of.children, A, downWeights)
}
\arguments{
\item{g}{graph (a Directed Acyclic Graph)}

\item{freqs}{observed genotype frequencies}

\item{no.of.children}{number of children for each node}

\item{A}{adjacency matrix of G}

\item{downWeights}{Down weights as computed by computeDWNW}
}
\value{
a vector containing the normalized Down weights for each edge
}
\description{
Normalizes down weights so that the sum of weights of edges exiting a node is 1
}
\examples{
normalizeUPW(g, freqs, no.of.children, A, downWeights)

}
