% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{normalizeUPW}
\alias{normalizeUPW}
\title{Up weights normalization}
\usage{
normalizeUPW(g, freqs, no.of.children, A, upWeights)
}
\arguments{
\item{g}{graph (a Directed Acyclic Graph)}

\item{freqs}{observed genotype frequencies}

\item{no.of.children}{number of children for each node}

\item{A}{adjacency matrix of G}

\item{upWeights}{Up weights as computed by computeUPW}
}
\value{
a vector containing the normalized Up weights for each edge
}
\description{
Normalizes up weights so that the sum of weights of edges entering in a node is 1
}
\examples{
normalizeUPW(g, freqs, no.of.children, A, upWeights)

}
