% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetGenerator.R
\name{plot_generator}
\alias{plot_generator}
\title{Plot a generator}
\usage{
plot_generator(generator)
}
\arguments{
\item{generator}{a generator}
}
\value{
a basic plot of this generator
}
\description{
Simple ggraph interface to draw a generator
}
\examples{
require(dplyr)

example_dataset() \%>\%
 make_generator_stub() \%>\% 
 set_generator_edges(
   list(
    "D", "A, D", 1 , 
    "A", "A, D", 1 , 
    "A, D", "A, C, D", 1 , 
    "A, D", "A, B, D", 1 , 
    "Clonal", "D", 1 , 
    "Clonal", "A", 1 , 
    "D", "D", 1 , 
    "A", "A", 1 , 
    "A, D", "A, D", 1 , 
    "A, C, D", "A, C, D", 1 , 
    "A, B, D", "A, B, D", 1 , 
    "Clonal", "Clonal", 1 
  )) \%>\% 
  finalize_generator \%>\% 
  plot_generator
 
}
