% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{prepare.labels}
\alias{prepare.labels}
\title{Prepare node labels based on genotypes}
\usage{
prepare.labels(samples, genes)
}
\arguments{
\item{samples}{input dataset (mutational matrix) as matrix}

\item{genes}{list of gene names (in the columns' order)}
}
\value{
the computed edge list
}
\description{
Prepare node labels so that each node is labelled with a
comma separated list of the alterated genes representing its associated genotype.
}
\details{
Note that after this procedure the user is expected also to run fix.clonal.genotype
to also add the clonal genortype to the mutational matrix if it is not present.
}
\examples{
labels <- prepare.labels(samples, genes)

}
