% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetGenerator.R
\name{prepare_generator_edge_set_command}
\alias{prepare_generator_edge_set_command}
\title{Prepare a command to add edge weights to a generator}
\usage{
prepare_generator_edge_set_command(generator, by = "labels")
}
\arguments{
\item{generator}{a generator}

\item{by}{"labels" or "samples" to use gene labels or sample labels as references for
edge identifiers.}
}
\value{
NULL (the string with the function calls is printed on the stdout)
}
\description{
Prints a string in the form of the command that 
sets weights for all the edges of this generator.
}
\examples{
require(dplyr)
example_dataset() \%>\% 
  make_generator_stub() \%>\%
  prepare_generator_edge_set_command()

}
