% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{read.CAPRI}
\alias{read.CAPRI}
\title{Read a "CAPRI" file}
\usage{
read.CAPRI(filename)
}
\arguments{
\item{filename}{path to file}
}
\value{
the described mutational matrix as a dataframe
}
\description{
Read a "CAPRI" formatted file from the file system
}
\examples{
read.CAPRI("pathToDataset/myDataset.CAPRI")

}
