% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{read}
\alias{read}
\title{Read a "CAPRI" file}
\usage{
read(filename)
}
\arguments{
\item{filename}{path to file}
}
\value{
the described mutational matrix as a dataframe
}
\description{
Read a "CAPRI" formatted file, as read.CAPRI
}
\examples{
read("pathToDataset/myDataset.CAPRI")

}
