% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_CAPRIpop_string}
\alias{read_CAPRIpop_string}
\title{Read "CAPRIpop" file from a string}
\usage{
read_CAPRIpop_string(txt)
}
\arguments{
\item{txt}{string in valid "CAPRIpop" format}
}
\value{
the described mutational matrix as a (sparse) matrix
}
\description{
Read a "CAPRIpop" formatted file, from a text string
}
\examples{
read_CAPRIpop_string("
s\\\\g A B C D freqs
S1 0 0 0 1 0.1
S2 1 0 0 0 0.1
S3 1 0 0 0 0.2
S4 1 0 0 1 0.3
S5 1 1 0 1 0.05
S6 1 1 0 1 0.1
S7 1 0 1 1 0.05
S8 1 1 0 1 0.01
")

}
