% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{remove_transitive_edges}
\alias{remove_transitive_edges}
\title{Remove transitive edges from an edgelist}
\usage{
remove_transitive_edges(E)
}
\arguments{
\item{E}{edge list, built from "build_topology_subset"}
}
\value{
a new edgelist without transitive edges (as a N*2 matrix)
}
\description{
Remove transitive edges from an edgelist. This procedure is temporary to
cover a bug in 'relations' package.
}
\examples{
l <- list(c(1,2),c(2,3), c(1,3))
remove_transitive_edges(l)

}
