% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{sample.mutations.hist}
\alias{sample.mutations.hist}
\title{Histogram of samples' frequencies}
\usage{
sample.mutations.hist(df, binwidth = 1)
}
\arguments{
\item{df}{input dataframe (mutational matrix)}

\item{binwidth}{binwidth parameter for the histogram (as in ggplot)}
}
\description{
Create the histogram of the samples' mutational frequencies
}
\examples{
sample.mutations.hist(df, binwidth = 10)

}
