% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{select.genes.on.mutations}
\alias{select.genes.on.mutations}
\title{Filter dataset by genes' mutation count}
\usage{
select.genes.on.mutations(df, n, desc = T)
}
\arguments{
\item{df}{input dataset (mutational matrix) to be reduced}

\item{n}{number of genes to be kept}

\item{desc}{T: select the n least mutated genes, F: select the n most mutated genes}
}
\value{
the modified dataset (mutational matrix)
}
\description{
Dataset filtering on genes, based on their mutation count
}
\examples{
# keep information on the 100 most mutated genes
select.genes.on.mutations(df, 100)
# keep information on the 100 least mutated genes
select.genes.on.mutations(df, 100, desc = F)

}
