% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{select.samples.on.mutations}
\alias{select.samples.on.mutations}
\title{Filter dataset by samples' mutation count}
\usage{
select.samples.on.mutations(df, n, desc = T)
}
\arguments{
\item{df}{input dataset (mutational matrix) to be reduced}

\item{n}{number of samples to be kept}

\item{desc}{T: select the n least mutated samples, F: select the n most mutated samples}
}
\value{
the modified dataset (mutational matrix)
}
\description{
Dataset filtering on samples, based on their mutation count
}
\examples{
# keep information on the 100 most mutated samples
select.samples.on.mutations(df, 100)
# keep information on the 100 least mutated samples
select.samples.on.mutations(df, 100, desc = F)
# combine selections
select.samples.on.mutations(df , 100, desc = F) \%>\%
    select.genes.on.mutations(100)

}
