% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{to.dot}
\alias{to.dot}
\title{Dot graph output}
\usage{
to.dot(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
a string representing the graph in dot format
}
\description{
Represents this graph in dot format (a textual output format)
}
\examples{
require(dplyr)
require(purrr)
preproc <- example.dataset() \%>\% dataset.preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph.non.transitive.subset.topology(samples, labels)
W <- compute.weights.default(g, freqs)
to.dot(g,W,labels)

}
