% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{update.df}
\alias{update.df}
\title{Dataset line by line construction: add a sample}
\usage{
\method{update}{df}(d, sampleName, ...)
}
\arguments{
\item{d}{an existing dataframe (mutational matrix)}

\item{sampleName}{the row (sample) name}

\item{...}{sample's genotype (0/1 numbers)}
}
\value{
the modified dataframe (mutational matrix)
}
\description{
Add a sample (a row) to an existing dataset. This procedure is meant to be used with the "%>%" operator
}
\examples{
make.dataset(APC,P53,KRAS)   \%>\%
    update.df("S1", 1, 0, 1) \%>\%
    update.df("S2", 1, 1, 1)

}
