% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMICE.R
\docType{package}
\name{CIMICE}
\alias{CIMICE}
\title{CIMICE Package}
\description{
R implementation of the CIMICE tool
}
\details{
CIMICE-R: (Markov) Chain Method to Inferr Cancer Evolution
}
\author{
Nicolò Rossi \email{olocin.issor@gmail.com}
}
