% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{computeDWNW.aux}
\alias{computeDWNW.aux}
\title{Up weights computation (aux)}
\usage{
computeDWNW.aux(g, edge, freqs, no.of.children, A, normUpWeights)
}
\arguments{
\item{g}{graph (a Directed Acyclic Graph)}

\item{edge}{the currently considered edge}

\item{freqs}{observed genotype frequencies}

\item{no.of.children}{number of children for each node}

\item{A}{adjacency matrix of G}

\item{normUpWeights}{normalized up weights as computed by normalizeUPW}
}
\value{
a vector containing the Up weights for each edge
}
\description{
Computes the Down weights formula:
}
\details{
W_{down}(<a,b>) = NormalizedW_{up}(<a,b>)(P(b) + \Sigma{for: x in L_b}{of: W_{down}(<b,x>)}

using a Dinamic Programming approach (recursion), see vignettes for further explaination.
}
\examples{
computeUPW.aux(g, edge, freqs, no.of.children, A)

}
