% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw.networkD3}
\alias{draw.networkD3}
\title{NetworkD3 graph output}
\usage{
draw.networkD3(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
networkD3 object representing g as described
}
\description{
Draws the output graph using networkD3
}
\examples{
require(dplyr)
preproc <- example.dataset() \%>\% dataset.preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph.non.transitive.subset.topology(samples, labels)
W <- compute.weights.default(g, freqs)
draw.networkD3(g,W,labels)

}
